move asset claim fetch to header
