+++ app/src/common/constants.ts
@@ -112,3 +112,6 @@ export const XDAI_NETWORKS = ['100', '0x64', '77', '0x4d']
 
 export const XDAI_TO_DAI_TOKEN_BRIDGE_ADDRESS = '0x7301CFA0e1756B71869E93d4e4Dca5c7d0eb0AA6'
 export const DAI_TO_XDAI_TOKEN_BRIDGE_ADDRESS = '0x4aa42145Aa6Ebf72e164C9bBC74fbD3788045016'
+
+export const RELAY_FEE = '1000000000000000' // 0.001 xdai
+export const RELAY_ADDRESS = '0xa493f3Adf76560092088a61e9e314a08D0B1B2b8'
